+++ exportmulti.go
@@ -4,6 +4,7 @@ import (
 	"archive/zip"
 	"bufio"
 	"io"
+	"log"
 	"math/rand"
 	"os"
 	"strconv"
@@ -19,14 +20,18 @@ func RandStringRunes(n int) string {
 	}
 	return string(b)
 }
-func ExportMultisheet(filename string, fetcher SheetFetcher) {
+func ExportMultisheet(filename string, fetcher SheetFetcher) error {
 	now := time.Now()
 	sheetName := now.Format("20060102150405") //filename should be (pseudo)random
-	shaStr, _ := os.Create(sheetName + ".ss")
+	shaStr, e := os.Create(sheetName + ".ss")
+	if e != nil {
+		log.Println(e)
+		return e
+	}
 	//defer shaStr.Close()
-	SharedStrWriter := bufio.NewWriter(shaStr)
-	SharedStrWriter.WriteString("<?xml version=\"1.0\" encoding=\"UTF-8\" standalone=\"yes\"?>")
-	SharedStrWriter.WriteString("<sst xmlns=\"http://schemas.openxmlformats.org/spreadsheetml/2006/main\" count=\"0\" uniqueCount=\"0\">")
+	sharedStrWriter := bufio.NewWriter(shaStr)
+	sharedStrWriter.WriteString("<?xml version=\"1.0\" encoding=\"UTF-8\" standalone=\"yes\"?>")
+	sharedStrWriter.WriteString("<sst xmlns=\"http://schemas.openxmlformats.org/spreadsheetml/2006/main\" count=\"0\" uniqueCount=\"0\">")
 	cellCount := 0
 	outputFile := filename
 	sheetNames := fetcher.GetSheetNames()
@@ -52,22 +57,42 @@ func ExportMultisheet(filename string, fetcher SheetFetcher) {
 			break
 		}
 		sheetName := now.Format("20060102150405") + RandStringRunes(5)
-		ExportWorksheet(sheetName, curSheet, SharedStrWriter, &cellCount)
+		e = ExportWorksheet(sheetName, curSheet, sharedStrWriter, &cellCount)
+		if e != nil {
+			log.Println(e)
+			return e
+		}
 		file["xl/worksheets/sheet"+strconv.Itoa(sheetCounter)+".xml"], _ = os.Open(sheetName)
 		toCloseList = append(toCloseList, "sheet"+strconv.Itoa(sheetCounter))
 		toDeleteList = append(toDeleteList, sheetName)
 		sheetCounter++
 	}
-	SharedStrWriter.WriteString("</sst>")
-	SharedStrWriter.Flush()
+	sharedStrWriter.WriteString("</sst>")
+	sharedStrWriter.Flush()
 	shaStr.Close()
-	file["xl/sharedStrings.xml"], _ = os.Open(sheetName + ".ss")
-	of, _ := os.Create(outputFile)
+	file["xl/sharedStrings.xml"], e = os.Open(sheetName + ".ss")
+	if e != nil {
+		log.Println(e)
+		return e
+	}
+	of, e := os.Create(outputFile)
+	if e != nil {
+		log.Println(e)
+		return e
+	}
 	defer of.Close()
 	zipWriter := zip.NewWriter(of)
 	for k, v := range file {
-		fWriter, _ := zipWriter.Create(k)
-		io.Copy(fWriter, v)
+		fWriter, e := zipWriter.Create(k)
+		if e != nil {
+			log.Println(e)
+			return e
+		}
+		_, e = io.Copy(fWriter, v)
+		if e != nil {
+			log.Println(e)
+			return e
+		}
 	}
 	for _, toClose := range toCloseList {
 		(file["xl/worksheets/"+toClose+".xml"]).(*os.File).Close()
@@ -75,7 +100,16 @@ func ExportMultisheet(filename string, fetcher SheetFetcher) {
 	for _, toDelete := range toDeleteList {
 		os.Remove("./" + toDelete)
 	}
-	zipWriter.Close()
+	e = zipWriter.Close()
+	if e != nil {
+		log.Println(e)
+		return e
+	}
 	(file["xl/sharedStrings.xml"]).(*os.File).Close()
-	os.Remove(sheetName + ".ss")
+	e = os.Remove(sheetName + ".ss")
+	if e != nil {
+		log.Println(e)
+		return e
+	}
+	return nil
 }
